+++ RELEASES.md
@@ -9,6 +9,7 @@ Breaking changes
 
 * `AST::iter_functions` now returns an iterator instead of taking a closure.
 * `Module::iter_script_fn_info` is removed and merged into `Module::iter_script_fn`.
+* `Module::get_script_function_by_signature` renamed to `Module::get_script_fn` and returns `&<Shared<ScriptFnDef>>`.
 * The `merge_namespaces` parameter to `Module::eval_ast_as_new` is removed and now defaults to `true`.
 * `GlobalFileModuleResolver` is removed because its performance gain over the `FileModuleResolver` is no longer very significant.
 * The following `EvalAltResult` variants are removed and merged into `EvalAltResult::ErrorMismatchDataType`: `ErrorCharMismatch`, `ErrorNumericIndexExpr`, `ErrorStringIndexExpr`, `ErrorImportExpr`, `ErrorLogicGuard`, `ErrorBooleanArgMismatch`
@@ -18,7 +19,6 @@ New features
 
 * `OptimizationLevel::Simple` now eagerly evaluates built-in binary operators of primary types (if not overloaded).
 * Added `is_def_var()` to detect if variable is defined, and `is_def_fn()` to detect if script function is defined.
-* Added `Module::get_script_fn` to get a scripted function in a module, if any, based on name and number of parameters.
 
 
 Version 0.19.0